adding lognormal
